requirements.txt
flask_sqlalchemy
flask_migrate
flask_wtf
flask_bcrypt
bleach
python-dotenv
